Update sources to compile with Rust 0.11

+++ main.rs
@@ -2,8 +2,9 @@
 #![no_std]
 #![feature(macro_rules)]
 #![feature(globs)]
+#![feature(lang_items)]
 
-#![crate_id="blinky#0.3.1"]
+#![crate_id="blinky#0.3.2"]
 
 use zero::std_types::*;
 use libarm::stm32f4xx::*;
@@ -20,12 +21,12 @@ mod libarm {
 	pub mod stm32f4xx_rcc;
 }
 
-static LED_GREEN :u32 = 12;
-static LED_ORANGE :u32 = 13;
-static LED_RED :u32 = 14;
-static LED_BLUE :u32 = 15;
+static LED_GREEN :uint = 12;
+static LED_ORANGE :uint = 13;
+static LED_RED :uint = 14;
+static LED_BLUE :uint = 15;
 
-static LED :u32 = LED_RED;
+static LED :uint = LED_RED;
 
 #[no_mangle]
 pub extern "C" fn TIM2_IRQHandler() {
@@ -54,7 +55,7 @@ pub extern "C" fn main()
 	let speed = GPIO_Speed_100MHz!() << (pin * 2);
 	let otype = GPIO_OType_PP!() << pin;
 	let pullup = GPIO_PuPd_NOPULL!() << (pin * 2);
-	let irq_en = 1 << (TIM2_IRQn!());
+	let irq_en = 1 << (TIM2_IRQn!() as uint);
 
 	RCC.AHB1ENR |= RCC_AHB1ENR_GPIODEN!(); // enable the clock to GPIOD
 	RCC.APB1ENR |= RCC_APB1ENR_TIM2EN!(); // enable TIM2 clock
